Working now.

+++ server/src/main/java/groupxii/database/Database.java
@@ -158,7 +158,7 @@ public class Database extends Thread {
 
     /** Finds an UserEntry by id.
      */
-    public DBObject findDocumentById(long id) {
+    public DBObject findDocumentById(int id) {
         BasicDBObject query = new BasicDBObject();
         query.put("userId", id);
         DBObject dbObject = userCollection.findOne(query);
@@ -179,7 +179,7 @@ public class Database extends Thread {
 
     /** receives two id's and adds the first one as a friend to the first one.
      */
-    public void addFriendId(long id1,long id2) {
+    public void addFriendId(int id1,int id2) {
         BasicDBObject newDocument = new BasicDBObject();
         newDocument.append("$addToSet", new BasicDBObject().append("friendsId", id1));
         BasicDBObject searchQuery = new BasicDBObject().append("userId", id2);
